+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/task/runTaskTool.ts
@@ -3,19 +3,19 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { timeout } from '../../../../../base/common/async.js';
-import { CancellationToken } from '../../../../../base/common/cancellation.js';
-import { localize } from '../../../../../nls.js';
-import { ITelemetryService } from '../../../../../platform/telemetry/common/telemetry.js';
-import { IChatService } from '../../../chat/common/chatService.js';
-import { ILanguageModelsService } from '../../../chat/common/languageModels.js';
-import { CountTokensCallback, IPreparedToolInvocation, IToolData, IToolImpl, IToolInvocation, IToolInvocationPreparationContext, IToolResult, ToolDataSource, ToolProgress } from '../../../chat/common/languageModelToolsService.js';
-import { ITaskService, ITaskSummary, Task } from '../../../tasks/common/taskService.js';
-import { ITerminalService } from '../../../terminal/browser/terminal.js';
-import { pollForOutputAndIdle, promptForMorePolling } from './bufferOutputPolling.js';
-import { getOutput } from './outputHelpers.js';
+import { timeout } from '../../../../../../base/common/async.js';
+import { CancellationToken } from '../../../../../../base/common/cancellation.js';
+import { localize } from '../../../../../../nls.js';
+import { ITelemetryService } from '../../../../../../platform/telemetry/common/telemetry.js';
+import { IChatService } from '../../../../chat/common/chatService.js';
+import { ILanguageModelsService } from '../../../../chat/common/languageModels.js';
+import { CountTokensCallback, IPreparedToolInvocation, IToolData, IToolImpl, IToolInvocation, IToolInvocationPreparationContext, IToolResult, ToolDataSource, ToolProgress } from '../../../../chat/common/languageModelToolsService.js';
+import { ITaskService, ITaskSummary, Task } from '../../../../tasks/common/taskService.js';
+import { ITerminalService } from '../../../../terminal/browser/terminal.js';
+import { pollForOutputAndIdle, promptForMorePolling } from '../bufferOutputPolling.js';
+import { getOutput } from '../outputHelpers.js';
 import { getTaskDefinition, getTaskForTool } from './taskHelpers.js';
-import { MarkdownString } from '../../../../../base/common/htmlContent.js';
+import { MarkdownString } from '../../../../../../base/common/htmlContent.js';
 
 type RunTaskToolClassification = {
 	taskId: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The ID of the task.' };
