+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/task/taskHelpers.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IStringDictionary } from '../../../../../base/common/collections.js';
-import { ConfiguringTask, Task } from '../../../tasks/common/tasks.js';
-import { ITaskService } from '../../../tasks/common/taskService.js';
+import { IStringDictionary } from '../../../../../../base/common/collections.js';
+import { ConfiguringTask, Task } from '../../../../tasks/common/tasks.js';
+import { ITaskService } from '../../../../tasks/common/taskService.js';
 
 
 export function getTaskDefinition(id: string) {
